+++ flutter_inappwebview/lib/src/in_app_webview/in_app_webview.dart
@@ -97,6 +97,8 @@ class InAppWebView extends StatefulWidget {
       int markTimeMillis,
     )?
     onPerformanceMarkMillis,
+    FutureOr<List<WebUri>?> Function(InAppWebViewController controller)?
+    onRequestVisitedHistory,
     void Function(InAppWebViewController controller, String? title)?
     onTitleChanged,
     FutureOr<AjaxRequestAction?> Function(
@@ -512,6 +514,9 @@ class InAppWebView extends StatefulWidget {
                        markName,
                        markTimeMillis,
                      )
+               : null,
+           onRequestVisitedHistory: onRequestVisitedHistory != null
+               ? (controller) => onRequestVisitedHistory.call(controller)
                : null,
            onTitleChanged: onTitleChanged != null
                ? (controller, title) => onTitleChanged.call(controller, title)