+++ flutter_inappwebview_platform_interface/lib/src/in_app_webview/platform_headless_in_app_webview.dart
@@ -94,6 +94,7 @@ class PlatformHeadlessInAppWebViewCreationParams
     super.onFirstContentfulPaintMillis,
     super.onLargestContentfulPaintMillis,
     super.onPerformanceMarkMillis,
+    super.onRequestVisitedHistory,
     super.onContentSizeChanged,
     super.onShowFileChooser,
     super.onInsertInputSuggestion,