+++ flutter_inappwebview_platform_interface/lib/src/in_app_webview/platform_inappwebview_widget.dart
@@ -102,6 +102,7 @@ class PlatformInAppWebViewWidgetCreationParams
     super.onFirstContentfulPaintMillis,
     super.onLargestContentfulPaintMillis,
     super.onPerformanceMarkMillis,
+    super.onRequestVisitedHistory,
     super.onContentSizeChanged,
     super.onShowFileChooser,
     super.onInsertInputSuggestion,